requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
scikit-learn==0.22.1
pandas==1.0.0
azureml-core==1.53.0
azureml-opendatasets==1.53.0
pyyaml
